first steps toward linting adjustements
