+++ .github/workflows/linting.yml
@@ -24,4 +24,4 @@ jobs:
         pip install pylint
     - name: Lint with pylint
       run: |
-        pylint exts/*.py
+        pylint exts/*.py executioner.py
