+++ exts/embedding.py
