+++ exts/slapping.py
